requirements.txt
streamlit
langchain
langchain-ollama
langchain-openai
python-dotenv
ollama
